+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.test.js
@@ -16,8 +16,10 @@ import uuidv4 from '../../global/js/utils/uuidv4';
 import { SidePanel } from '.';
 import { Add16 } from '@carbon/icons-react';
 
-const blockClass = `${pkg.prefix}--side-panel`;
-const actionSetBlockClass = `${pkg.prefix}--action-set`;
+const { devtoolsAttribute, getDevtoolsId, prefix } = pkg;
+
+const blockClass = `${prefix}--side-panel`;
+const actionSetBlockClass = `${prefix}--action-set`;
 const sizes = ['xs', 'sm', 'md', 'lg', 'max'];
 
 const dataTestId = uuidv4();
@@ -28,6 +30,7 @@ const pageContentSelectorValue = '#side-panel-test-page-content';
 
 const onRequestCloseFn = jest.fn();
 const onUnmountFn = jest.fn();
+
 const renderSidePanel = ({ ...rest }, children = <p>test</p>) =>
   render(
     <SidePanel
@@ -36,7 +39,8 @@ const renderSidePanel = ({ ...rest }, children = <p>test</p>) =>
         open: true,
         onRequestClose: onRequestCloseFn,
         ...rest,
-      }}>
+      }}
+    >
       {children}
     </SidePanel>
   );
@@ -68,7 +72,8 @@ const SlideIn = ({
         }
         placement={placement}
         onUnmount={onUnmountFn}
-        {...rest}>
+        {...rest}
+      >
         Content
       </SidePanel>
       <div id={pageContentSelectorValue.slice(1)} />
@@ -207,7 +212,8 @@ describe('SidePanel', () => {
         title={title}
         includeOverlay
         open={false}
-        onRequestClose={onRequestCloseFn}>
+        onRequestClose={onRequestCloseFn}
+      >
         Content
       </SidePanel>
     );
@@ -388,6 +394,15 @@ describe('SidePanel', () => {
     expect(ref.current).toEqual(screen.getByRole('complementary'));
   });
 
+  it('adds the Devtools attribute to the containing node', () => {
+    renderSidePanel({ 'data-testid': dataTestId });
+
+    expect(screen.getByTestId(dataTestId)).toHaveAttribute(
+      devtoolsAttribute,
+      getDevtoolsId(SidePanel.displayName)
+    );
+  });
+
   it('should call the onRequestClose event handler', () => {
     const { click } = userEvent;
     const { container } = renderSidePanel();
@@ -492,7 +507,8 @@ describe('SidePanel', () => {
           labelText={labelText}
           placement="left"
           open
-          usePageContentSelector>
+          usePageContentSelector
+        >
           Content
         </SlideIn>
       );